+++ js/src/index.js
@@ -32,7 +32,6 @@ const AXES_PADDING_X = 60;
 const AXES_PADDING_X_WITH_LABEL = AXES_PADDING_X + AXES_LABEL_SIZE;
 const AXES_PADDING_Y = 20;
 const AXES_PADDING_Y_WITH_LABEL = AXES_PADDING_Y + AXES_LABEL_SIZE;
-const TOOLTIP_EVENT_TYPE = 'tooltip';
 const TOOLTIP_DEBOUNCE_TIME = 250;
 const TOOLTIP_MANDATORY_VISUAL_PROPERTIES = (/** @type {const} */ ({ x: 'X', y: 'Y' }));
 const TOOLTIP_OPTIONAL_VISUAL_PROPERTIES = (/** @type {const} */ ({ color: 'Col', opacity: 'Opa', size: 'Size' }));
@@ -97,7 +96,6 @@ const properties = {
   connectionSize: 'pointConnectionSize',
   connectionSizeBy: 'pointConnectionSizeBy',
   viewDownload: 'viewDownload',
-  viewReset: 'viewReset',
   viewSync: 'viewSync',
   hovering: 'hovering',
   axes: 'axes',
@@ -207,6 +205,7 @@ class JupyterScatterView {
   constructor({ el, model }) {
     this.el = el;
     this.model = model;
+    this.eventTypes = model.get('event_types');
   }
 
   render() {
@@ -390,9 +389,31 @@ class JupyterScatterView {
   }
 
   customEventHandler(event) {
-    if (event.type === TOOLTIP_EVENT_TYPE) {
-      if (event.index !== this.tooltipPointIdx) return;
+    if (event.type === this.eventTypes.TOOLTIP) {
+      if (event.index !== this.tooltipPointIdx && event.show !== true) return;
       this.tooltipDataHandlers(event)
+      if (event.show) this.showTooltip(event.index);
+    }
+    if (event.type === this.eventTypes.VIEW_RESET) {
+      if (!this.scatterplot) return;
+      if (event.area) {
+        this.scatterplot.zoomToArea(
+          event.area,
+          {
+            transition: event.animation > 0,
+            transitionDuration: event.animation,
+            transitionEasing: 'quadInOut',
+          }
+        );
+      } else {
+        this.scatterplot.zoomToOrigin(
+          {
+            transition: event.animation > 0,
+            transitionDuration: event.animation,
+            transitionEasing: 'quadInOut',
+          }
+        );
+      }
     }
   }
 
@@ -1506,7 +1527,7 @@ class JupyterScatterView {
 
     this.tooltipContentsUpdater = (pointIdx) => {
       this.model.send({
-        type: TOOLTIP_EVENT_TYPE,
+        type: this.eventTypes.TOOLTIP,
         index: pointIdx,
         properties: this.tooltipPropertiesNonVisual,
         preview: this.tooltipPreview,
@@ -2014,7 +2035,7 @@ class JupyterScatterView {
     if (newPoints.length === this.scatterplot.get('points').length) {
       // We assume point correspondence
       this.scatterplot.draw(newPoints, {
-        transition: true,
+        transition: this.model.get('transition_points'),
         transitionDuration: 3000,
         transitionEasing: 'quadInOut',
         preventFilterReset: this.model.get('prevent_filter_reset'),
@@ -2342,14 +2363,6 @@ class JupyterScatterView {
     });
   }
 
-  viewResetHandler() {
-    this.scatterplot.reset();
-    setTimeout(() => {
-      this.model.set('view_reset', false);
-      this.model.save_changes();
-    }, 0);
-  }
-
   optionsHandler(newOptions) {
     this.scatterplot.set(newOptions);
   }
@@ -2404,4 +2417,5 @@ async function render({ model, el }) {
   view.render();
   return () => view.destroy();
 }
-export default {render}
+
+export default { render };